Changed IP client detection for Apache 2.4
-Using the updated 2.4 client IP address accessor per Apache documentation as referenced below.  In cases where mods like mod_remoteip are used this method will take advantage of the updated address set by mod_remoteip.  For other configurations the connected client ip is still used. 

This change allows configurations that use mod_remoteip to set the client IP from a proxy header and treat it as if it is the client ip throughout the rest of processing.  In such a configuration IP2LocationDetectProxy can be On or Off.  In a configuration without this change the usage of mod_remoteip and its behavior either have to be disabled (affecting other things) or altered such that a new header is injected for this mod to find the IP in question.

-Added IP2LOCATION_IP env/note variable that will be set to the address used to perform the lookup.  This is useful in debugging and possibly logging.

https://httpd.apache.org/docs/2.4/ru/developer/new_api_2_4.html
---
conn_rec->remote_ip and conn_rec->remote_addr
These fields have been renamed in order to distinguish between the client IP address of the connection and the useragent IP address of the request (potentially overridden by a load balancer or proxy). References to either of these fields must be updated with one of the following options, as appropriate for the module:
-When you require the IP address of the user agent, which might be connected directly to the server, or might optionally be separated from the server by a transparent load balancer or proxy, use request_rec->useragent_ip and request_rec->useragent_addr.
-When you require the IP address of the client that is connected directly to the server, which might be the useragent or might be the load balancer or proxy itself, use conn_rec->client_ip and conn_rec->client_addr.
---